kata: Cleanup and moving types to Models folder.
