+++ ConAppCommandPattern/Model/ConcreteCommand.cs
@@ -1,0 +1,9 @@
+﻿namespace ConAppCommandPattern.Model;
+
+public class ConcreteCommand(Receiver receiver) : Command(receiver ?? new Receiver())
+{
+    public override void Execute()
+    {
+        Receiver.Action();
+    }
+}
