+++ ConAppCommandPattern/Model/Receiver.cs
@@ -1,0 +1,11 @@
+﻿using static System.Console;
+
+namespace ConAppCommandPattern.Model;
+
+public class Receiver
+{
+    public static void Action()
+    {
+        WriteLine("Receiver.Action() called");
+    }
+}
